Created GIAB Report
Now makes a raw data dump that can be fed into a specific google sheet
to create a formatted, fleshed out report.

Updated the README to reflect changes and some spelling corrections.
Keeping the make_tinydb and giab_report for notes. But I'll be deleting
it before merging this branch back in
Small change to multi_venn