better vcf handling and new giabreport anno
We no longer try to test 'fetch' on the comparison vcf to prevent
some errors. We now just check for the .gz and .tbi.
New TR INFO tags in the GIAB truth set are parsed for giabreport
When the SVTYPE or END flags are not "Number=1", we now handle
those lists values by picking the first one. However, the more
correct behavior would be to handle multi-allelic sites.