AF fix and collapse bugs
Collapse --hap wasn't working correctly. The assumptions about the calls
being phased wasn't being properly used (don't collapse 1|1) and
the NumCollapsed was being populated before the single-best match was chosen.
The latter is a reporting problem, but the former had an effect on the
results with ~3% of collapsed calls being mis-collapsed

af_calc had a syntax error