GRM bed, repmask speed, consistency report
GRM now supports a bed file, which will allow one to make the parallel
ranges over the bed regions instead of the input --reference contigs.
This is very useful when variants were called in a subset of chromosomes
and therefore other chromosomes' tabix fetch would return null. This can
represent a significant speed improvement

repmask default parameters to RepeatMasker now includes -qq, which
presumably run faster (at a sensitivity cost)

consistency's report for PctOfFileCalls was ordered poorly. The report
is now accurate with the same order as the KEY/list of files input