Auto-versioning changes
When truvari is installed from the repo and if it's a develop version, add the commit hash to the version.
If the repo has unstaged changes, 'uc' is appended to the version as well.
This enables better tracking when results are created off of develop.
setup_logging now optionally shows this version information at the head of logs. This is turned on for
all commands that use setup_logging but will remain off by default for outside tools/scripts that reuse setup_logging