found rebench bug
When msa2vcf fails to make an output, phab catches and ignores the error. Then rebench was looking for phab's output which doesn't exist. This caused the region to fail. Now we check that the output is made. If not, an informative warning is raised and the region's original counts are kept.