storing debugging
It isn't fixed but I need to save notes.
I think that Docker is having a hard time handling the SIMD instructions
of pyabpoa. This is causing some of the larger sequences to get stick
(or perhaps die) during run_poa. There are no signals the task is gone
so the main just hangs forever. I suspect Docker because everything
works just fine outside of it.